Moving Files and refactoring code
