+++ app/src/main/AndroidManifest.xml
@@ -27,7 +27,7 @@
         android:icon="@drawable/icon"
         android:label="@string/app_name" >
         <service
-            android:name=".Main_Service"
+            android:name=".Services.Main_Service"
             android:enabled="true"
             android:label="@string/serviceLabel"
             android:permission="android.permission.BIND_WALLPAPER" >
@@ -41,16 +41,16 @@
         </service>
 
         <activity
-            android:name=".SettingsActivity"
+            android:name=".Activities.SettingsActivity"
             android:label="@string/preferenceLabel" >
         </activity>
         <activity
-            android:name=".ImageZoneSelectionActivity"
+            android:name=".Activities.ImageZoneSelectionActivity"
             android:label="@string/title_activity_image_zone_selection"
-            android:parentActivityName=".SettingsActivity" >
+            android:parentActivityName=".Activities.SettingsActivity" >
             <meta-data
                 android:name="android.support.PARENT_ACTIVITY"
-                android:value=".SettingsActivity" />
+                android:value=".Activities.SettingsActivity" />
         </activity>
     </application>
